Zenyx - TL/requirements.txt
# Telegram Bot API
python-telegram-bot==20.8

# Variáveis de ambiente
python-dotenv==1.0.0

# Redis para persistência
redis==5.0.1

# Requisições HTTP
requests==2.31.0

# Manipulação de datas
python-dateutil==2.8.2

# Geração de QR Code (caso necessário)
qrcode[pil]==7.4.2

# Logging avançado
loguru==0.7.2

# Utilitários
typing-extensions==4.9.0

# Async utilities
aiohttp==3.9.1

# JSON handling
ujson==5.9.0

# Validação de dados
pydantic==2.5.3

# Timezone
pytz==2023.3.post1

# Criptografia (para tokens seguros)
cryptography==41.0.7

# Geração de códigos únicos
shortuuid==1.0.11

# Agendamento de tarefas
apscheduler==3.10.4

# Monitoring (opcional)
psutil==5.9.8

# Database migrations (se necessário)
alembic==1.13.1

# Tests
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# Development
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0